% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risk2.R
\name{compute_hum_loop}
\alias{compute_hum_loop}
\title{Compute IPCW-adjusted Hypervolume Under the Manifold}
\usage{
compute_hum_loop(outcome_mat, pred_mat, weight_mat, tie_correction = TRUE)
}
\arguments{
\item{outcome_mat}{Output from get_outcome_mat function}

\item{pred_mat}{Output from calc_risks function}
}
\value{
a scalar HUM
}
\description{
This function computes the IPCW-adjusted Hypervolume under the manifold, as in Lee (unpublished).
}
