% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqID_HReg2.R
\name{FreqID_HReg2}
\alias{FreqID_HReg2}
\title{Fit Parametric Frailty Illness-Death Model for Semi-Competing Risks Data}
\usage{
FreqID_HReg2(
  Formula,
  data,
  na.action = "na.fail",
  subset = NULL,
  hazard = c("weibull"),
  frailty = TRUE,
  model,
  knots_list = NULL,
  nP0 = rep(4, 3),
  startVals = NULL,
  hessian = TRUE,
  quad_method = "kronrod",
  n_quad = 15,
  optim_method = "BFGS",
  control = NULL
)
}
\arguments{
\item{Formula}{a Formula object, with the outcome on the left of a
\code{~}, and covariates on the right. It is of the form, \code{time to non-terminal
  event + corresponding censoring indicator | time to terminal event
  + corresponding censoring indicator ~ covariates for h_1 |
  covariates for h_2 | covariates for h_3}. For example, \code{y_1 + delta_1 | y_2 + delta_2 ~ x_1 | x_2 | x_3}.}

\item{data}{a \code{data.frame} in which to interpret the variables named in Formula.}

\item{na.action}{how NAs are treated. See \code{\link[stats]{model.frame}}.}

\item{subset}{a specification of the rows to be used: defaults to all rows. See \code{\link[stats]{model.frame}}.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{frailty}{Boolean indicating whether a gamma distributed subject-specific frailty should
be included. Currently this must be set to \code{TRUE}.}

\item{model}{String (not case sensitive) specifying the transition assumption: either \code{"semi-Markov"}}

\item{knots_list}{Used for hazard specifications besides Weibull, a
list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard. If
\code{NULL}, will be created by \code{\link{get_default_knots_list}}.}

\item{nP0}{vector of length three of integers indicating how many baseline hazard parameters
should be specified for each of the three transition hazards. This input is only relevant when
hazard is something other than "weibull" and is superceded by knots_list.}

\item{startVals}{A numeric vector of parameter starting values, arranged as follows:
the first \eqn{k_1+k_2+k_3} elements correspond to the baseline hazard parameters,
then the \eqn{k_1+k_2+k_3+1} element corresponds to the gamma frailty log-variance parameter,
then the last\eqn{q_1+q_2+q_3} elements correspond with the regression parameters.
If set to \code{NULL}, will be generated automatically using \code{\link{get_start}}.}

\item{hessian}{Boolean indicating whether the hessian (aka, the inverse of the covariance matrix)
should be computed and returned.}

\item{quad_method}{String indicating which quadrature method to use to evaluate numerical integral of B-spline.
Options are 'kronrod' for Gauss-Kronrod quadrature or 'legendre' for Gauss-Legendre quadrature.}

\item{n_quad}{Scalar for number of Gaussian quadrature points used to evaluate numerical integral of B-spline.}

\item{optim_method}{a string naming which \code{optim} method should be used.}

\item{control}{a list of control attributes passed directly into the \code{optim} function.}
}
\value{
\code{FreqID_HReg2} returns an object of class \code{Freq_HReg}.
}
\description{
Fit Parametric Frailty Illness-Death Model for Semi-Competing Risks Data
}
