% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_selection.R
\name{FreqID_HReg2_forward}
\alias{FreqID_HReg2_forward}
\title{Forward Selection Procedure for Parametric Illness-Death Model}
\usage{
FreqID_HReg2_forward(
  time1_name,
  time2_name,
  event1_name,
  event2_name,
  trunc_name = NULL,
  vars,
  data,
  na.action = "na.fail",
  subset = NULL,
  weights = NULL,
  hazard = c("weibull"),
  frailty = TRUE,
  model,
  knots_list = NULL,
  fixed1 = character(0),
  fixed2 = character(0),
  fixed3 = character(0),
  k = 2,
  steps = 1000,
  control = NULL,
  verbose = 1,
  quad_method = "kronrod",
  n_quad = 15,
  optim_method = "BFGS",
  extra_starts = 0,
  output_options = NULL
)
}
\arguments{
\item{vars}{character string with all of the possible variable names to be searched through.}

\item{data}{a \code{data.frame} in which to interpret
the variables named in Formula.}

\item{na.action}{how NAs are treated. See \code{\link[stats]{model.frame}}.}

\item{subset}{a specification of the rows to be used: defaults to all rows.
See \code{\link[stats]{model.frame}}.}

\item{weights}{Vector of numeric weights of length \eqn{n} corresponding
to each observed individual.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{frailty}{Boolean indicating whether a gamma distributed subject-specific frailty should
be included. Currently this must be set to \code{TRUE}.}

\item{model}{String (not case sensitive) specifying the transition assumption: either \code{"semi-Markov"}}

\item{knots_list}{Used for hazard specifications besides Weibull, a
list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard.
If \code{NULL}, will be created by \code{\link{get_default_knots_list}}
according to the number of parameters specified by \code{nP0}.}

\item{fixed1}{a string vector for parameters you want guaranteed included in arm 1}

\item{fixed2}{a string vector for parameters you want guaranteed included in arm 2}

\item{fixed3}{a string vector for parameters you want guaranteed included in arm 3}

\item{k}{the multiple of the number of degrees of freedom used for the penalty. Only k = 2 gives the genuine AIC: k = log(n) is sometimes referred to as BIC or SBC.}

\item{steps}{the maximum number of steps to be considered. The default is 1000 (essentially as many as required). It is typically used to stop the process early.}

\item{control}{a list of control attributes passed directly
into the \code{optim} function.}

\item{verbose}{if positive, information is printed during the running of step. Larger values may give more detailed information.}

\item{quad_method}{String indicating which quadrature method to use to
evaluate numerical integral of B-spline. Options are
\code{"kronrod"} for Gauss-Kronrod quadrature or
\code{"legendre"} for Gauss-Legendre quadrature.}

\item{n_quad}{Scalar for number of Gaussian quadrature points used to
evaluate numerical integral of B-spline.}

\item{optim_method}{a string naming which \code{optim}
method should be used for optimization.}

\item{extra_starts}{Integer giving the number of extra
starts to try when optimizing.}

\item{output_options}{List of named boolean elements specifying whether
certain additional components should be included in
the model object output. Options include
\itemize{
\item{\code{Finv} Variance-covariance matrix. Defaults to \code{TRUE}.}
\item{\code{fit_nf} If \code{frailty} is \code{TRUE}, include an
additional fit object for the corresponding non-frailty model.
Defaults to \code{TRUE}.}
\item{\code{eb_frailties} Include vector of empirical Bayes
predicted gamma frailty values. Defaults to \code{TRUE}.}
\item{\code{grad_mat} Matrix with rowwise
score vectors for each individual evaluated at the MLE.
Used to compute "cheese" or "meat" in robust standard error computation.
Defaults to \code{FALSE}.}
\item{\code{cheese} Sum of outer products of individual score vectors,
used as the "cheese" or "meat" in robust standard error computation.
Defaults to \code{TRUE}.}
\item{\code{data} Original data frame used to fit model.
Defaults to \code{FALSE}.}
}}
}
\value{
A list.
}
\description{
This method is meant as a comparator for the penalized estimation procedure.
}
