% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_basis}
\alias{get_basis}
\title{Get Basis Function Values for Flexible Hazard Specifications}
\usage{
get_basis(y, knots_vec, hazard, deriv = FALSE)
}
\arguments{
\item{y}{Numeric vector of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{knots_vec}{Used for hazard specifications besides Weibull, a
vector increasing sequence of integers, each corresponding to
the knots for the flexible model on the baseline hazard.
If \code{NULL}, will be created by \code{\link{get_default_knots}}
according to the number of parameters specified by \code{nP0}.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{deriv}{Boolean for whether returned values should be from derivatives of
basis functions if \code{TRUE}, or original basis functions if \code{FALSE}.}
}
\value{
A matrix with each row corresponding to an element of the input, and
each column giving the corresponding basis function value.
}
\description{
Generate a matrix of basis functions evaluated at each value of a vector \code{x}.
}
