% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_default_knots}
\alias{get_default_knots}
\title{Generate Knot Location Vector from Event Times}
\usage{
get_default_knots(y, delta, p0, hazard)
}
\arguments{
\item{y}{Numeric vector of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta}{Numeric vector of length \eqn{n}  with indicator of 1 if the event was observed and 0 otherwise}

\item{p0}{Integer indicating how many baseline hazard parameters
should be specified for each of the three transition hazards.
This input is only relevant when hazard is something other than
\code{"weibull"} and is superceded by knots_vec.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}
}
\value{
A vector of increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard.
For cubic B-splines on log-hazard: min 0, max is max observed value of \code{y},
\code{p0-4} internal knots placed at quantiles of observed events \code{y[delta==1]}.
For Royston-Parmar (restricted cubic spline on log-cumulative hazard): min and max at min and max observed event times,
\code{p0-2} internal knots placed at quantiles of observed events
For piecewise constant: min 0, remaining \code{p0-1} knots placed at quantiles of observed events.
}
\description{
This function creates a list containing three numeric vectors. Each numeric vector
is a sequence of increasing integers giving the location of knots used for
spline and piecewise baseline hazard specifications. This function
generates this list according to the conventions and recommended locations
of these knots, which depends on the choice of hazard specification, number
of knots requested, and distribution of observed event times.
}
