% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_default_knots_list}
\alias{get_default_knots_list}
\title{Generate List of Knot Location Vectors from Event Times}
\usage{
get_default_knots_list(y1, y2, delta1, delta2, p01, p02, p03, hazard, model)
}
\arguments{
\item{y1}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{y2}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta1}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{delta2}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{p01, p02, p03}{Integers indicating how many parameters the model for each
transition baseline hazard should specify.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{model}{String (not case sensitive) specifying the transition assumption: either \code{"semi-Markov"}}
}
\value{
A list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard.
}
\description{
This function creates a list containing three numeric vectors of increasing integers giving the location of knots used for
spline and piecewise baseline hazard specifications, corresponding with the
illness-death model transition hazards. The knots are generated according to
the description of \code{\link{get_default_knots}}.
}
