% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_fit_nf}
\alias{get_fit_nf}
\title{Get Non-Frailty Illness-Death Model Fit}
\usage{
get_fit_nf(
  startVals_nf,
  y1,
  y2,
  delta1,
  delta2,
  Xmat1,
  Xmat2,
  Xmat3,
  knots_list,
  basis1,
  basis2,
  basis3,
  basis3_y1,
  dbasis1,
  dbasis2,
  dbasis3,
  hazard,
  model,
  control,
  hessian,
  optim_method,
  extra_starts
)
}
\arguments{
\item{startVals_nf}{A vector of starting parameter values without a frailty variance,
such as computed by \code{\link{get_start}}.}

\item{y1}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{y2}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta1}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{delta2}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{Xmat1}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{Xmat2}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{Xmat3}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{knots_list}{Used for hazard specifications besides Weibull, a
list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard. If
\code{NULL}, will be created by \code{\link{get_default_knots_list}}.}

\item{basis1}{Numeric matrices with \eqn{n} rows and \eqn{k_1,k_2,k_3} columns
with piecewise/spline basis function values at the corresponding \code{y1} and \code{y2} values.
Under semi-Markov model, basis3 represents basis derived from \eqn{y_2-y_1} and \code{basis3_y1} is unused,
while under Markov model, basis3 represents basis derived from \eqn{y_2} and \code{basis3_y1} is from \eqn{y_1}
Not used under Weibull model.}

\item{basis2}{Numeric matrices with \eqn{n} rows and \eqn{k_1,k_2,k_3} columns
with piecewise/spline basis function values at the corresponding \code{y1} and \code{y2} values.
Under semi-Markov model, basis3 represents basis derived from \eqn{y_2-y_1} and \code{basis3_y1} is unused,
while under Markov model, basis3 represents basis derived from \eqn{y_2} and \code{basis3_y1} is from \eqn{y_1}
Not used under Weibull model.}

\item{basis3}{Numeric matrices with \eqn{n} rows and \eqn{k_1,k_2,k_3} columns
with piecewise/spline basis function values at the corresponding \code{y1} and \code{y2} values.
Under semi-Markov model, basis3 represents basis derived from \eqn{y_2-y_1} and \code{basis3_y1} is unused,
while under Markov model, basis3 represents basis derived from \eqn{y_2} and \code{basis3_y1} is from \eqn{y_1}
Not used under Weibull model.}

\item{basis3_y1}{Numeric matrices with \eqn{n} rows and \eqn{k_1,k_2,k_3} columns
with piecewise/spline basis function values at the corresponding \code{y1} and \code{y2} values.
Under semi-Markov model, basis3 represents basis derived from \eqn{y_2-y_1} and \code{basis3_y1} is unused,
while under Markov model, basis3 represents basis derived from \eqn{y_2} and \code{basis3_y1} is from \eqn{y_1}
Not used under Weibull model.}

\item{dbasis1}{Numeric matrices with \eqn{n} rows and \eqn{k_1,k_2,k_3} columns
with piecewise/spline basis function derivative values at the corresponding \code{y1} and \code{y2} values.
Used only under Royston-Parmar model.}

\item{dbasis2}{Numeric matrices with \eqn{n} rows and \eqn{k_1,k_2,k_3} columns
with piecewise/spline basis function derivative values at the corresponding \code{y1} and \code{y2} values.
Used only under Royston-Parmar model.}

\item{dbasis3}{Numeric matrices with \eqn{n} rows and \eqn{k_1,k_2,k_3} columns
with piecewise/spline basis function derivative values at the corresponding \code{y1} and \code{y2} values.
Used only under Royston-Parmar model.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{model}{String (not case sensitive) specifying the transition assumption: either \code{"semi-Markov"}}

\item{control}{a list of control attributes passed directly into the \code{optim} function.}

\item{hessian}{Boolean indicating whether the hessian (aka, the inverse of the covariance matrix)
should be computed and returned.}

\item{optim_method}{a string naming which \code{optim} method should be used.}

\item{extra_starts}{Integer giving the number of extra starts to try when optimizing.}
}
\value{
A vector of starting parameter values.
}
\description{
A fitting function used to estimate a non-frailty illness-death model.
}
