% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_fit_uni}
\alias{get_fit_uni}
\title{Fit Univariate Model}
\usage{
get_fit_uni(
  startVals,
  y,
  delta,
  yL,
  anyLT,
  Xmat,
  basis,
  dbasis,
  basis_yL,
  hazard,
  control,
  hessian,
  optim_method,
  extra_starts
)
}
\arguments{
\item{startVals}{A numeric vector of parameter starting values, arranged as follows:
the first \eqn{k_1+k_2+k_3} elements correspond to the baseline hazard parameters,
then the last\eqn{q_1+q_2+q_3} elements correspond with the regression parameters.
If set to \code{NULL}, will be generated automatically using \code{\link{get_start}}.}

\item{y}{Numeric vector of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta}{Numeric vector of length \eqn{n}  with indicator of 1 if the event was observed and 0 otherwise}

\item{yL}{Numeric vector of length \eqn{n} with left-truncation times.}

\item{anyLT}{Boolean (or 1/0) value that is \code{TRUE} (or 1) if there are any non-zero left truncation times, and \code{FALSE} (or 0) otherwise}

\item{Xmat}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{basis}{Numeric matrix with \eqn{n} rows and \eqn{k} columns
with piecewise/spline basis function values at the corresponding \code{y} values.
Not used under Weibull model.}

\item{dbasis}{Numeric matrix with \eqn{n} rows and \eqn{k} columns
with piecewise/spline basis function derivative values at the corresponding \code{y} values.
Used only under Royston-Parmar model.}

\item{basis_yL}{Numeric matrix with \eqn{n} rows and \eqn{k} columns
with piecewise/spline basis function values at the corresponding \code{yL} values.
Not used under Weibull model.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{control}{a list of control attributes passed directly into the \code{optim} function.}

\item{hessian}{Boolean indicating whether the hessian (aka, the inverse of the covariance matrix)
should be computed and returned.}

\item{optim_method}{a string naming which \code{optim} method should be used for optimization.}

\item{extra_starts}{Integer giving the number of extra starts to try when optimizing.}
}
\value{
A fit object.
}
\description{
A function to fit.
}
