% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risk2.R
\name{get_ipcw_list}
\alias{get_ipcw_list}
\title{Get inverse probability of censoring weights}
\usage{
get_ipcw_list(
  yL,
  y2,
  delta2,
  t_cutoff,
  prec = 1e-05,
  prec2 = 100 * .Machine$double.eps
)
}
\arguments{
\item{yL}{Numeric vector of length \eqn{n} with left-truncation times.}

\item{t_cutoff}{Numeric vector indicating the time(s) to compute the risk profile.}
}
\value{
a vector.
}
\description{
This function returns a vector of inverse probability of censoring weights from an unadjusted Cox model
for censoring times.
}
