% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_quad_pointsweights}
\alias{get_quad_pointsweights}
\title{Function to return quadrature points and weights}
\usage{
get_quad_pointsweights(n_quad, quad_method)
}
\arguments{
\item{n_quad}{Scalar for number of Gaussian quadrature points used to evaluate numerical integral of B-spline.}

\item{quad_method}{String indicating which quadrature method to use to evaluate numerical integral of B-spline.
Options are 'kronrod' for Gauss-Kronrod quadrature or 'legendre' for Gauss-Legendre quadrature.}
}
\value{
A list with two named elements (\code{points} and \code{weights}), each
of which is a numeric vector with length equal to the number of
quadrature nodes
}
\description{
This function stores quadrature points used for numerical integration by either
Gauss-Kronrod or Gauss-Legendre rules. The values are pre-computed and stored for
number of nodes equal to 7, 11, 15, 21, 31, 41, 51, or 61.
}
