% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_start_uni}
\alias{get_start_uni}
\title{Get Starting Parameter Values for Univariate Model}
\usage{
get_start_uni(
  y,
  delta,
  yL,
  anyLT,
  Xmat,
  knots,
  basis,
  hazard,
  sparse_start = FALSE
)
}
\arguments{
\item{y}{Numeric vector of length \eqn{n} with (possibly censored) event times}

\item{delta}{Numeric vector of length \eqn{n}  with indicator of 1 if the event was observed and 0 otherwise}

\item{Xmat}{Numeric matrix with \eqn{n} rows and \eqn{q} columns containing covariates.}

\item{knots}{Increasing vector of integers corresponding to the knots used
in the desired spline or piecewise specification. Often an element of
list generated from \code{\link{get_default_knots_list}}.}

\item{basis}{Numeric matrix with \eqn{n} rows and \eqn{k} columns
with piecewise/spline basis function values at the corresponding \code{y} values.}

\item{hazard}{String specifying the form of the baseline hazard.}

\item{sparse_start}{Boolean indicating whether to set all regression parameters
to 0 if \code{TRUE}, or to pre-estimate them using univariate models if \code{FALSE}.}

\item{dbasis}{Numeric matrices with \eqn{n} rows and \eqn{k} columns
with piecewise/spline basis function derivative values at the corresponding \code{y} values.
Used only under Royston-Parmar model.}
}
\value{
A vector of starting parameter values.
}
\description{
A function to generate principled starting values for optimization, based on
model specifications.
}
