% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_ngrad_nhess_functions.R
\name{nll_ngrad_uni_func}
\alias{nll_ngrad_uni_func}
\title{Combined function with Negative Log-Likelihood Function and Gradient for Univariate Survival Model}
\usage{
nll_ngrad_uni_func(
  para,
  y,
  delta,
  yL,
  anyLT,
  Xmat,
  hazard,
  weights,
  basis,
  dbasis,
  basis_yL
)
}
\arguments{
\item{para}{A numeric vector of parameters, arranged as follows:
the first \eqn{k} elements correspond to the baseline hazard parameters,
then the last\eqn{q} elements correspond with the regression parameters.}

\item{y}{Numeric vector of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta}{Numeric vector of length \eqn{n}  with indicator of 1 if the event was observed and 0 otherwise}

\item{yL}{Numeric vector of length \eqn{n} with left-truncation times.}

\item{anyLT}{Boolean (or 1/0) value that is \code{TRUE} (or 1) if there are any non-zero left truncation times, and \code{FALSE} (or 0) otherwise}

\item{Xmat}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{weights}{Vector of numeric weights of length \eqn{n} corresponding
to each observed individual.}

\item{basis}{Numeric matrix with \eqn{n} rows and \eqn{k} columns
with piecewise/spline basis function values at the corresponding \code{y} values.
Not used under Weibull model.}

\item{dbasis}{Numeric matrix with \eqn{n} rows and \eqn{k} columns
with piecewise/spline basis function derivative values at the corresponding \code{y} values.
Used only under Royston-Parmar model.}

\item{basis_yL}{Numeric matrix with \eqn{n} rows and \eqn{k} columns
with piecewise/spline basis function values at the corresponding \code{yL} values.
Not used under Weibull model.}
}
\value{
Returns numeric sum of negative log likelihood contributions.
}
\description{
Function returning the negative log-likelihood for the illness-death model,
under specified baseline hazard, and specified frailty,
and specified Markov/semi-Markov transition assumption.
Typically, this function will not be used directly by the user, but as part of a
larger estimation procedure.
}
