% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_frailties.R
\name{pred_frailties}
\alias{pred_frailties}
\title{Predict Gamma Frailties}
\usage{
pred_frailties(
  para,
  y1,
  delta1,
  y2,
  delta2,
  yL = NULL,
  Xmat1,
  Xmat2,
  Xmat3,
  hazard,
  knots_list = NULL,
  model,
  quad_method = "legendre",
  n_quad = 15
)
}
\arguments{
\item{para}{A numeric vector of parameters, arranged as follows:
the first \eqn{k_1+k_2+k_3} elements correspond to the baseline hazard parameters,
then the \eqn{k_1+k_2+k_3+1} element corresponds to the gamma frailty log-variance parameter,
then the last\eqn{q_1+q_2+q_3} elements correspond with the regression parameters.}

\item{y1, y2}{Numeric vectors of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta1, delta2}{Numeric vectors of length \eqn{n}  with indicators of 1 if the event was observed and 0 otherwise}

\item{yL}{Numeric vector of length \eqn{n} with left-truncation times.}

\item{Xmat1, Xmat2, Xmat3}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{knots_list}{Used for hazard specifications besides Weibull, a
list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard.
If \code{NULL}, will be created by \code{\link{get_default_knots_list}}
according to the number of parameters specified by \code{nP0}.}

\item{model}{String (not case sensitive) specifying the transition assumption: either \code{"semi-Markov"}}

\item{quad_method}{String indicating which quadrature method to use to
evaluate numerical integral of B-spline. Options are
\code{"kronrod"} for Gauss-Kronrod quadrature or
\code{"legendre"} for Gauss-Legendre quadrature.}

\item{n_quad}{Scalar for number of Gaussian quadrature points used to
evaluate numerical integral of B-spline.}
}
\value{
if Xmat has only one row, and t_cutoff is a scalar, then returns a 1 by 4 element row matrix
of the following values:
\itemize{
\item{\code{mean} of subject's gamma empirical bayes posterior, i.e., shape/rate.)}
\item{\code{median} of subject's gamma empirical bayes posterior.)}
\item{\code{shape} parameter (as used in, e.g., \code{dgamma}, of subject's gamma empirical bayes posterior.)}
\item{\code{rate} parameter (as used in, e.g., \code{dgamma}, of subject's gamma empirical bayes posterior.)}
}
If Xmat has \code{n} rows, then returns an \code{n} by 4 matrix.
If Xmat has \code{n} rows and t_cutoff is a vector of length \code{s}, then returns an \code{s} by 4 by \code{n} array.
}
\description{
This function predicts estimated gamma frailties via empirical bayes
}
