% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrData.R
\docType{data}
\name{scrData}
\alias{scrData}
\title{A simulated clustered semi-competing risks data set}
\format{
\subsection{\code{scrData}}{

A data frame with 2000 rows and 8 columns:
\describe{
\item{time1}{the time to non-terminal event}
\item{event1}{the censoring indicators for the
non-terminal event time; 1=event observed, 0=censored/truncated}
\item{time2}{the time to terminal event}
\item{event2}{the censoring indicators for the
terminal event time; 1=event observed, 0=censored}
\item{cluster}{cluster numbers}
\item{x1,x2,x3}{vectors of continuous covarates}
}
}
}
\usage{
scrData
}
\description{
Simulated semi-competing risks data
}
\keyword{datasets}
