% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summary.R
\name{vcov_helper}
\alias{vcov_helper}
\title{Compute estimates of variance-covariance matrix}
\usage{
vcov_helper(
  Finv = NULL,
  cheese = NULL,
  n,
  name_vec = NULL,
  var_type = c("modelbased", "sandwich", "outer"),
  df_adjust = FALSE
)
}
\arguments{
\item{Finv}{inverse of observed information matrix}

\item{cheese}{summed score function outer products}

\item{n}{number of observations}

\item{name_vec}{String vector giving the desired names of each row/column of the matrix}

\item{var_type}{String giving type of variance covariance matrix estimate to return.
Options are "modelbased" which returns the inverse of the observed information,
"outer" which returns the inverse of the summed score function outer products,
or "sandwich" which returns the sandwich estimator}

\item{df_adjust}{Boolean for whether to include "finite sample"
degrees-of-freedom adjustment by multiplication with \eqn{n/(n-k)}
where \eqn{n} is the number of observations and
\eqn{k} is the number of estimated parameters.}
}
\description{
This function computes different variance estimate from frequentist
model object, using the observed information and/or the summed outer products
of each individual subject's scores evaluated at the MLE.
}
