% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov_sandwich}
\alias{vcov_sandwich}
\title{Extract robust "sandwich" estimate of variance-covariance matrix}
\usage{
vcov_sandwich(object, adjust = FALSE, ...)
}
\arguments{
\item{object}{Model object}

\item{adjust}{Boolean for whether to include "finite sample"
degrees-of-freedom adjustment by multiplication with \eqn{n/(n-k)}
where \eqn{n} is the number of observations and
\eqn{k} is the number of estimated parameters.}

\item{...}{Other arguments (currently unused).}
}
\description{
This function computes the robust "sandwich" variance estimate from frequentist
model object, provided that the model object includes the "cheese" or "meat" matrix comprising
the summed outer products of each individual subject's scores evaluated at the MLE.
}
