% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqID_HReg2.R
\name{FreqID_HReg2}
\alias{FreqID_HReg2}
\title{Fit Parametric Frailty Illness-Death Model for Semi-Competing Risks Data}
\usage{
FreqID_HReg2(
  Formula,
  data,
  na.action = "na.fail",
  subset = NULL,
  weights = NULL,
  hazard = c("weibull"),
  frailty = TRUE,
  model,
  knots_list = NULL,
  nP0 = rep(4, 3),
  startVals = NULL,
  control = NULL,
  quad_method = "kronrod",
  n_quad = 15,
  optim_method = "BFGS",
  extra_starts = 0,
  output_options = NULL
)
}
\arguments{
\item{Formula}{a Formula object, with the outcome on the left of a
\code{~}, and covariates on the right. It is of the form,
\code{left truncation time | time to non-terminal
  event + corresponding censoring indicator | time to terminal event
  + corresponding censoring indicator ~ covariates for h_1 |
  covariates for h_2 | covariates for h_3}.
For example, \code{y_L | y_1 + delta_1 | y_2 + delta_2 ~ x_1 | x_2 | x_3}.
If there is no left truncation, then the lefthand side should only contain
the two sets of outcome variables.}

\item{data}{a \code{data.frame} in which to interpret
the variables named in Formula.}

\item{na.action}{how NAs are treated. See \code{\link[stats]{model.frame}}.}

\item{subset}{a specification of the rows to be used: defaults to all rows.
See \code{\link[stats]{model.frame}}.}

\item{weights}{Vector of numeric weights of length \eqn{n} corresponding
to each observed individual.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{frailty}{Boolean indicating whether a gamma distributed subject-specific frailty should
be included. Currently this must be set to \code{TRUE}.}

\item{model}{String (not case sensitive) specifying the transition assumption: either \code{"semi-Markov"}}

\item{knots_list}{Used for hazard specifications besides Weibull, a
list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard.
If \code{NULL}, will be created by \code{\link{get_default_knots_list}}
according to the number of parameters specified by \code{nP0}.}

\item{nP0}{vector of length three of integers indicating how many
baseline hazard parameters
should be specified for each of the three transition hazards.
This input is only relevant when
hazard is something other than "weibull" and is superceded by knots_list.}

\item{startVals}{A numeric vector of parameter starting values,
arranged as follows:
the first \eqn{k_1+k_2+k_3} elements are the baseline hazard parameters,
then the gamma frailty log-variance parameter,
then the last\eqn{q_1+q_2+q_3} elements are the regression parameters.
If set to \code{NULL}, generated internally using \code{\link{get_start}}.}

\item{control}{a list of control attributes passed directly
into the \code{optim} function.}

\item{quad_method}{String indicating which quadrature method to use to
evaluate numerical integral of B-spline. Options are
\code{"kronrod"} for Gauss-Kronrod quadrature or
\code{"legendre"} for Gauss-Legendre quadrature.}

\item{n_quad}{Scalar for number of Gaussian quadrature points used to
evaluate numerical integral of B-spline.}

\item{optim_method}{a string naming which \code{optim}
method should be used for optimization.}

\item{extra_starts}{Integer giving the number of extra
starts to try when optimizing.}

\item{output_options}{List of named boolean elements specifying whether
certain additional components should be included in
the model object output. Options include
\itemize{
\item{\code{Finv} Variance-covariance matrix. Defaults to \code{TRUE}.}
\item{\code{fit_nf} If \code{frailty} is \code{TRUE}, include an
additional fit object for the corresponding non-frailty model.
Defaults to \code{TRUE}.}
\item{\code{eb_frailties} Include vector of empirical Bayes
predicted gamma frailty values. Defaults to \code{TRUE}.}
\item{\code{grad_mat} Matrix with rowwise
score vectors for each individual evaluated at the MLE.
Used to compute "cheese" or "meat" in robust standard error computation.
Defaults to \code{FALSE}.}
\item{\code{beta_test} Matrix in "long" form giving Likelihood Ratio Test
and Score Test statistics for each regression coefficient. Defaults to \code{TRUE}.}
\item{\code{cheese} Sum of outer products of individual score vectors,
used as the "cheese" or "meat" in robust standard error computation.
Defaults to \code{TRUE}.}
\item{\code{data} Original data frame used to fit model.
Defaults to \code{FALSE}.}
}}
}
\value{
\code{FreqID_HReg2} returns an object of class \code{Freq_HReg}.
}
\description{
Fit Parametric Frailty Illness-Death Model for Semi-Competing Risks Data
}
\examples{
#loading a data set
data(scrData)

#fitting Weibull semi-Markov illness-death model with gamma frailties
form <- Formula::Formula(time1 + event1 | time2 + event2 ~ x1 + x2 + x3 | x1 + x2 | x1 + x2)
fit_WB	<- FreqID_HReg2(Formula = form, data=scrData, model="semi-Markov",
extra_starts = 0,hazard = "weibull",frailty = TRUE,optim_method = c("BFGS"))

#exploring results
fit_WB
summ.fit_WB <- summary(fit_WB); names(summ.fit_WB)
summ.fit_WB
pred_WB <- predict(fit_WB, tseq=seq(from=0.1, to=30, length.out=100))
plot(pred_WB, plot.est="Haz")
plot(pred_WB, plot.est="Surv")

}
