% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_default_knots}
\alias{get_default_knots}
\title{Generate Knot Location Vector from Event Times}
\usage{
get_default_knots(y, delta, p0, hazard)
}
\arguments{
\item{y}{Numeric vector of length \eqn{n} with (possibly censored) event times}

\item{delta}{Numeric vector of length \eqn{n}  with indicator of 1 if the event was observed and 0 otherwise}

\item{p0}{Integer indicating how many parameters the model for the baseline hazard should specify.}

\item{hazard}{String specifying the form of the baseline hazard.}
}
\value{
A list of three increasing sequences of integers, each corresponding to
the knots for the flexible model on the corresponding transition baseline hazard.
}
\description{
This function creates a list containing three numeric vectors. Each numeric vector
is a sequence of increasing integers giving the location of knots used for
spline and piecewise baseline hazard specifications. This function
generates this list according to the conventions and recommended locations
of these knots, which depends on the choice of hazard specification, number
of knots requested, and distribution of observed event times.
}
