% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helper_functions.R
\name{get_start_uni}
\alias{get_start_uni}
\title{Get Starting Parameter Values for Univariate Model}
\usage{
get_start_uni(
  y,
  delta,
  yL,
  anyLT,
  Xmat,
  knots_vec,
  basis,
  hazard,
  weights,
  sparse_start = FALSE
)
}
\arguments{
\item{y}{Numeric vector of length \eqn{n} with (possibly censored) non-terminal and terminal event times}

\item{delta}{Numeric vector of length \eqn{n}  with indicator of 1 if the event was observed and 0 otherwise}

\item{yL}{Numeric vector of length \eqn{n} with left-truncation times.}

\item{anyLT}{Boolean (or 1/0) value that is \code{TRUE} (or 1) if there are any non-zero left truncation times, and \code{FALSE} (or 0) otherwise}

\item{Xmat}{Numeric matrices with \eqn{n} rows and \eqn{q_1,q_2,q_3} columns containing covariates.}

\item{knots_vec}{Used for hazard specifications besides Weibull, a
vector increasing sequence of integers, each corresponding to
the knots for the flexible model on the baseline hazard.
If \code{NULL}, will be created by \code{\link{get_default_knots}}
according to the number of parameters specified by \code{nP0}.}

\item{basis}{Numeric matrix with \eqn{n} rows and \eqn{k} columns
with piecewise/spline basis function values at the corresponding \code{y} values.
Not used under Weibull model.}

\item{hazard}{String (not case sensitive) specifying the form of the baseline hazard.
Options are \code{"weibull"} for Weibull,
\code{"piecewise"} for piecewise constant on the hazard scale,
\code{"bspline"} for cubic B-spline on the log-hazard scale,
and \code{"royston-parmar"} for restricted cubic spline on the log-cumulative hazard scale.
Aliases for these are \code{"wb"}, \code{"pw"}, \code{"bs"}, and \code{"rp"} respectively.}

\item{weights}{Vector of numeric weights of length \eqn{n} corresponding
to each observed individual.}

\item{sparse_start}{Boolean indicating whether to set all regression parameters
to 0 if \code{TRUE}, or to pre-estimate them using univariate models if \code{FALSE}.}
}
\value{
A vector of starting parameter values.
}
\description{
A function to generate principled starting values for optimization, based on
model specifications.
}
